backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
appdirs==1.4.4
argcomplete==3.6.3
argon2-cffi==23.1.0
argon2-cffi-bindings==25.1.0
attrs==25.4.0
azure-core==1.38.0
azure-storage-blob==12.28.0
bcrypt==5.0.0
bidict==0.23.1
boto3==1.42.29
botocore==1.42.29
brotli==1.2.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coverage==7.13.1
cryptography==43.0.1
cssselect2==0.8.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
et_xmlfile==2.0.0
fastapi==0.128.0
fonttools==4.61.1
frozenlist==1.8.0
google-api-core==2.29.0
google-auth==2.47.0
google-cloud-core==2.5.0
google-cloud-storage==3.8.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
gunicorn==25.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
importlib_resources==6.5.2
iniconfig==2.3.0
isodate==0.7.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.3
MarkupSafe==3.0.3
multidict==6.7.0
nltk==3.9.2
numpy==2.4.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
passlib==1.7.4
paypalrestsdk==1.13.3
pillow==12.1.0
playwright==1.57.0
pluggy==1.6.0
prometheus_client==0.24.1
propcache==0.4.1
proto-plus==1.27.0
protobuf==6.33.4
psutil==7.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydifact==0.1.8
pydyf==0.12.1
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.10.1
pyngrok==7.5.0
pyOpenSSL==25.1.0
pyotp==2.9.0
PyPDF2==3.0.1
pyphen==0.17.2
pyqtgraph==0.14.0
pyserial==3.5
PySide6==6.10.1
PySide6_Addons==6.10.1
PySide6_Essentials==6.10.1
pytest==9.0.2
pytest-cov==7.0.0
pytest-mock==3.15.1
pytest-qt==4.5.0
python-barcode==0.16.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-engineio==4.13.0
python-escpos==3.1
python-jose==3.5.0
python-multipart==0.0.21
python-socketio==5.16.0
pytz==2025.2
pyusb==1.3.1
PyYAML==6.0.3
qrcode==8.2
django-redis==5.4.0
hiredis==2.3.2
redis==5.0.1
regex==2026.1.15
reportlab==4.4.7
requests==2.32.5
rsa==4.9.1
s3transfer==0.16.0
schedule==1.2.2
scikit-learn==1.8.0
scipy==1.17.0
setuptools==80.9.0
shiboken6==6.10.1
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.45
sqlparse==0.5.5
starlette==0.50.0
stripe==14.1.0
textblob==0.19.0
threadpoolctl==3.6.0
tinycss2==1.5.1
tinyhtml5==2.0.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uvicorn==0.40.0
watchfiles==1.1.1
weasyprint==67.0
webencodings==0.5.1
websockets==15.0.1
whitenoise==6.11.0
wsproto==1.3.2
yarl==1.22.0
zopfli==0.4.0
psycopg2-binary==2.9.11
Django>=5.0
djangorestframework
celery
sentry-sdk
django-environ
django-cors-headers
drf-spectacular
djangorestframework-simplejwt
django-filter
channels
openai
structlog